extracted WorkspaceProcessor, moved FileHandler
