+++ javalab/src/main/java/com/smorales/javalab/middleware/config/JAXRSConfiguration.java
@@ -1,6 +1,6 @@
 package com.smorales.javalab.middleware.config;
 
-import com.smorales.javalab.middleware.buildtool.rest.ProcessorResource;
+import com.smorales.javalab.middleware.buildtool.rest.WorkspaceProcessorResource;
 
 import javax.ws.rs.core.Application;
 import java.util.Set;
@@ -16,7 +16,7 @@ public class JAXRSConfiguration extends Application {
     }
 
     private void addRestResourceClasses(Set<Class<?>> resources) {
-        resources.add(ProcessorResource.class);
+        resources.add(WorkspaceProcessorResource.class);
         resources.add(CORSFilter.class);
     }
 }
